Moved slice methods to `util` module.
