+++ tests/net_dynamic_hb.rs
@@ -4,48 +4,16 @@ extern crate rand;
 extern crate threshold_crypto;
 
 pub mod net;
+pub mod util;
 
-use std::{collections, ops};
+use std::collections;
 
 use hbbft::dynamic_honey_badger::{Change, ChangeState, DynamicHoneyBadger, Input};
 use hbbft::messaging::DistAlgorithm;
 
 use net::adversary::NullAdversary;
 use net::VirtualNet;
-
-trait SubSlice
-where
-    Self: ops::Index<ops::Range<usize>>,
-{
-    #[inline]
-    fn subslice(
-        &self,
-        range: ops::Range<usize>,
-    ) -> &<Self as ops::Index<ops::Range<usize>>>::Output;
-}
-
-impl<T> SubSlice for [T] {
-    #[inline]
-    fn subslice(
-        &self,
-        mut range: ops::Range<usize>,
-    ) -> &<Self as ops::Index<ops::Range<usize>>>::Output {
-        if range.end > self.len() {
-            range.end = self.len();
-        }
-
-        &self[range]
-    }
-}
-
-fn choose_approx<R: ?Sized + rand::Rng, T: Clone>(
-    rng: &mut R,
-    mut slice: &[T],
-    mut n: usize,
-    out_of_first: usize,
-) {
-    slice = &slice[..(out_of_first.min(slice.len()))];
-}
+use util::SubSlice;
 
 // FIXME: User better batch size, etc.
